refactor: extract :set parser and applier so non-interactive callers can reuse them
